Implement authentication and authorization
Users must be autheticated to delete or update their account. Currently
they must provide some number in place of user id. It doesn't have to be
their own id. But if this id happens to belong to another user the
delete or update action will not actually affect that user. It will
affect only the current authenticated user.