Add authorization on updating and destroying users
Previously user could try to update or delete another user. He would
fail of course, but the action was still performed on the user itself.
Now if tries to do something like this he will get an unauthorized
error. It is done with the help of `friendly_id` gem and adding a name
field.